+++ aws-sdk-core/lib/aws-sdk-core/xml/builder.rb
@@ -65,9 +65,14 @@ module Aws
         value_ref = ref.shape.value
         node(name, ref) do
           hash.each do |key, value|
-            node('entry', ref)  do
+            if ref['flattened']
               member(key_ref.location_name || 'key', key_ref, key)
               member(value_ref.location_name || 'value', value_ref, value)
+            else
+              node('entry', ref)  do
+                member(key_ref.location_name || 'key', key_ref, key)
+                member(value_ref.location_name || 'value', value_ref, value)
+              end
             end
           end
         end